Work In Progress - Lazy Data Loading
The loading bar is complete, and once a row item is hovered, it will
display the loading bar after 0.5 seconds.

TODO: After the loading is completed, once that same item is hovered
again, it should display the data, and not the loading again. This can
be based of a bool. Every rowItem has a "DataIsLoaded" bool, which will
be used to keep track of where to put the loading screen

TODO: Move as much of the logic as possible to the server side, instead
of (how it is now, temporarily) on the client side.