Major Improvement
Now properly displays different data for every person. I also moved most
of the logic to the server side.

Still a work in progress though.

TODO:
Sometimes both the loading and the data doesnt show (most likely a fault
in the logic).
+
At the moment they work by clicking (due to JSON not sending/observing
the correct changes on hover)
+
Clear out a lot of unused code and comments