Everything works as intended now
There is a list of people. Every person has a slot for "favorite game".
When a user hovers each person, a loading will start (at least
visually), and when it completes, it will have loaded the info (which
their favorite game is)

As this information is loaded, it is also filled into the rows.

When a user starts loading data, and then leaves the object (with the
mouse), that data needs to be re-loaded

TODO: Write meaningful text, perhaps change data "retrieving" method to
ScheduleTask, and perhaps simplify the code a bit